requirements.txt
openai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
hubspot-api-client
python-dotenv
flask
requests
psycopg2-binary
gunicorn
